refactor: update API route
- Ensured dynamic rendering is enforced for the invoice generation process.
